requirements.txt
python-dotenv
discord